worker/requirements.txt
yt-dlp
supabase
python-dotenv
requests
